poisson/requirements.txt
playwright==1.58.0
aiohttp==3.13.3
pyyaml==6.0.3
numpy==2.4.2
